src/threema_service.py
import os
import json
import logging
from threema.gateway import Connection, GatewayError, MessageError

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class ThreemaService:
    def __init__(self):
        # Using the provided credentials
        self.identity = os.getenv('THREEMA_IDENTITY', '*3MAGW01')
        self.secret = os.getenv('THREEMA_SECRET', '&YwrCeMju6ApTHNRpa6p')
        
        # Initialize connection
        try:
            self.connection = Connection(
                identity=self.identity,
                secret=self.secret
            )
            logger.info("Threema connection initialized")
        except Exception as e:
            logger.error(f"Failed to initialize Threema connection: {e}")
            self.connection = None
    
    def send_message(self, recipient, message):
        """
        Send a message to a Threema recipient
        
        Args:
            recipient (str): Threema ID of the recipient
            message (str): Message text to send
            
        Returns:
            dict: Result of the operation
        """
        if not self.connection:
            logger.warning("Threema connection not initialized")
            return {
                'success': False,
                'error': 'Threema connection not initialized'
            }
        
        try:
            # For this PoC, we'll simulate sending
            # In production, uncomment the following line:
            # message_id = self.connection.send_text(recipient, message)
            
            # Simulate successful sending
            logger.info(f"Simulated sending message to {recipient}: {message}")
            message_id = "simulated_message_id"
            
            return {
                'success': True,
                'message_id': message_id
            }
        except (GatewayError, MessageError) as e:
            logger.error(f"Failed to send Threema message: {e}")
            return {
                'success': False,
                'error': str(e)
            }
        except Exception as e:
            logger.error(f"Unexpected error sending Threema message: {e}")
            return {
                'success': False,
                'error': f"Unexpected error: {str(e)}"
            }

# Singleton instance
threema_service = ThreemaService()
